purge built-in Notification, the app must now provide it
Previously, there was an elaborate, custom framework for customizing the
Notification that was hard-coded into CacheWord.  Now, the app must supply
a Notification, and that triggers CacheWordService to run as a foreground
Service.  Otherwise, CacheWordService will run as a background Service with
no Notification.

This removes the last remnants that require CacheWord to be an Android
Library Project.  Now it can be distributed as just a jar.  It can still be
used as a Library Project for those who like doing it that way.

fixes #3711 https://dev.guardianproject.info/issues/3711
fixes #3680 https://dev.guardianproject.info/issues/3680